Avoid invalid instances to be saved
As the time is ticking, and I'll deliver this code tomorrow, I won't
spend much more time tryinh to figure it out why the "disabled" flag did
not have the expected effect, but I found a way to do in a more simplistic
fashion.

I'll keep this in mind, and I'll try to fix it as soon as possible by
adding "zod" package to perform a better validation with error messages
in the future.